requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
mysqlclient==2.2.7
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
